requirements.txt
discord.py-self
colorama==0.4.4
asyncio==3.4.3
